+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Major: Add item rarity to NPC loot notifications. (#425)
 - Minor: Add chat message upon noteworthy plugin updates. (#429)
+- Minor: Allow loot notifications for rare items below the min value threshold. (#426)
 - Minor: Include relevant kill count in collection log notifications. (#424)
 - Minor: Obtain kill count from chat commands plugin for loot notifications. (#392)
 - Minor: Add chat command to obtain the player's dink hash. (#408)